requirements.txt
fastapi
uvicorn
jinja2
pandas
python-dotenv
python-multipart
openai>=1.0.0